AZ布局不展示问题3

+++ services/service/src/device_manager_service.cpp
@@ -2325,14 +2325,17 @@ int32_t DeviceManagerService::GetTrustedDeviceList(const std::string &pkgName, s
     return DM_OK;
 }
 
-void DeviceManagerService::HandleDataChange(const char *peerUdid, const char *groupInfo)
+void DeviceManagerService::HandleDataChange(const char *peerUdid, const GroupInfo &groupInfo)
 {
     LOGI("DeviceManagerService::HandleDataChange start, peerUdid = %{public}s.", peerUdid);
-    std::vector<DistributedDeviceProfile::AccessControlProfile> profiles 
+    std::vector<DistributedDeviceProfile::AccessControlProfile> profiles
         = DeviceProfileConnector::GetInstance().GetAllAccessControlProfile();
     LOGI("profiles size = %{public}lu", profiles.size());
     for (auto &item : profiles) {
-        if (item.GetTrustDeviceId() == peerUdid) {
+        if (item.GetBindType() == groupInfo.groupType &&
+            item.GetTrustDeviceId() == peerUdid &&
+            item.GetAccesser().GetAccesserUserId() == groupInfo.userId &&
+            item.GetAccesser().GetAccesserAccountId() == groupInfo.osAccountId) {
             DeviceProfileConnector::GetInstance().DeleteAccessControlById(item.GetAccessControlId());
         }
     }
